+++ Cry3DEngine/terrain_sector_render.cpp
@@ -546,7 +546,7 @@ void CSectorInfo::RenderSector(CCObject * pTerrainCCObject)
 
 void CSectorInfo::AddLowResSectorIndex(int _x, int _y, int _step, int _lod)
 {
-  unsigned short id = _x/_step*(CTerrain::GetTerrainSize()/_step+1) + _y/_step;
+  unsigned int id = _x/_step*(CTerrain::GetTerrainSize()/_step+1) + _y/_step;
 
 	m_nLowResTerrainIdxRange[_lod][0] = min(m_nLowResTerrainIdxRange[_lod][0],id);
 	m_nLowResTerrainIdxRange[_lod][1] = max(m_nLowResTerrainIdxRange[_lod][1],id);
